requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
executing==1.2.0
idna==3.4
inflection==0.5.1
ipdb==0.13.11
ipython==8.10.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
